port various tweaks from trunk
git-svn-id: https://svn.r-project.org/R/branches/R-2-15-branch@59881 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/appl/machar.c
@@ -118,7 +118,7 @@ void machar(int *ibeta, int *it, int *irnd, int *ngrd, int *machep, int *negep,
 	do {
 		b = b + b;
 		temp = a + b;
-		itemp = temp - a;
+		itemp = (int)(temp - a);
 	}
 	while (itemp == 0);
 	*ibeta = itemp;
